Build user query more declaritively
Tbh idk if this is necessary. These abstractions make it
easier to declaratively build sql queries, but honestly I don't really
have that big of an issue manually writing them myself. I have some
boilerplate one way or another, whether it's a bunch of sql query
literals, or this query builder stuff. I initially began considering
having some sort of framework to make it easier to restrict users to
their own content, but I'm not so sure how necessary it is/if all this
extra code to build queries will be worth the slightly improved devx
when it comes to user's viewing context.