Implement module versions
Why is this needed?

Say you are a student and you're half way through a course.
The teacher updates a previous module you've completed,
changing one of the questions to be completely different,
along with other content.
As the student you go back to review that module. How do we render to
the user what choice they made? They can't review that content?
Or if a student is halfway through a module and it gets edited.
What gets rendered when they go the next block?

My solution for this is to just have different module versions,
and all the blocks/content/questions are tied to the version.
If a student starts/completes a module, we pin their view of the module
to that version, and only people seeing the module for the first time
see the latest version.