[db_migration] Move question and choice text to content
We want to allow markdown/math content in the question and choice text,
so we migrate the current string columns to be pointers to rows in the
content table. We could have just left things how they were and started
rendering the text as markdown/math, but we want to keep things in line
with we current avoid excessive content duplication between versions.

Unfortunately we cannot just add a column, backfill, and apply the
foreign key constraint in sqlite, so we have to create entirely new
tables for questions and choices.

Because answers and explanations also reference question as a foreign
key, we also gotta make new tables for those.