Fix content deletion queries
This was rough. If something had the same content as something else,
since we forgot to exclude content reference elsewhere (i.e. content
referenced in the just added module version), we would delete content.
Since content id is now a foreign key on question, this would also
delete the question!