Add note and deprecate latest migration code
I should figure out a better way of handling this.

I am changing some code that's used in the migration in an upcoming
commit, so I either need to refactor the migration code (useless, it's
already in prod), or get rid of it somehow. I'd like to keep a record of
the migrations for reference when I make future migrations, but maybe
that's not important.. For now I choose to leave things and just comment
out the code.